+++ NEWS.md
@@ -1,4 +1,14 @@
 
+# PathwayGuidedRF 0.3.1
+
+minor changes:
+- allow underscore in variable names
+- allow to specify number of cases and controls in function sim.data.study.2
+- return gene specific importance in function pw.rf.hunt
+  
+bug fixes:
+- enable one element vector of effects in function sim.data.study.2
+
 # PathwayGuidedRF 0.3.0
 
 initial commit to GitHub